JohnBryce-Lessons/2026.01.06/01 - ANG + FLASK/backend/run.py
"""
Application Entry Point.

This script creates and runs the Flask application.
Usage: python run.py
"""

import os
from dotenv import load_dotenv

# Load environment variables from .env file
load_dotenv()

from app import create_app

# Create application instance
app = create_app()

if __name__ == '__main__':
    host = os.getenv('FLASK_HOST', '0.0.0.0')
    port = int(os.getenv('FLASK_PORT', 5000))
    debug = os.getenv('FLASK_DEBUG', 'true').lower() == 'true'

    print(f"Starting Flask server on {host}:{port}")
    print(f"Debug mode: {debug}")
    print(f"Environment: {os.getenv('FLASK_ENV', 'dev')}")

    app.run(host=host, port=port, debug=debug)
