JohnBryce-Lessons/2026.01.06/01 - ANG + FLASK/backend/requirements.txt
# Flask framework
Flask==3.0.0

# Database
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5

# Authentication
Flask-JWT-Extended==4.6.0

# CORS support
Flask-CORS==4.0.0

# Environment variables
python-dotenv==1.0.0

# Database drivers
psycopg2-binary==2.9.9  # PostgreSQL
# pymysql==1.1.0          # MySQL (uncomment if needed)

# Development tools
# pytest==7.4.3
# pytest-cov==4.1.0
